Better project folders
- Create `/layout/master_detail` folder ;
- create `/layout/compact` folder ;
- create `/controls/progess_indicator` folder ;
- create `/controls/togglable` folder ;
- create `/controls/toggle_button` folder.
