+++ lib/src/controls/togglable/yaru_radio.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
-import '../constants.dart';
+import '../../constants.dart';
+import '../toggle_button/yaru_radio_button.dart';
 import 'yaru_checkbox.dart';
-import 'yaru_radio_button.dart';
 import 'yaru_switch.dart';
 import 'yaru_togglable.dart';
