+++ lib/src/controls/togglable/yaru_switch.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
+import '../toggle_button/yaru_switch_button.dart';
 import 'yaru_checkbox.dart';
 import 'yaru_radio.dart';
-import 'yaru_switch_button.dart';
 import 'yaru_togglable.dart';
 
 const _kSwitchActivableAreaPadding =
