+++ lib/src/controls/toggle_button/yaru_radio_button.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'yaru_radio.dart';
+import '../togglable/yaru_radio.dart';
 import 'yaru_toggle_button.dart';
 
 /// A desktop style radio button with an interactive label.
