+++ lib/src/layouts/master_detail/yaru_master_detail_theme.dart
@@ -2,7 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:yaru/yaru.dart';
 
-import '../constants.dart';
+import '../../constants.dart';
 
 /// Holds theme data for [YaruMasterDetailTheme].
 @immutable
